chore: reconcile the Gitea and GitHub mirrors of main (#34)
Reconciles the Gitea and GitHub mirrors of `main`, which have been unable to
fast-forward in either direction since 2026-08-15.

## Why they diverged

| Side | Commits it alone carried |
| --- | --- |
| Gitea `main` | aislop complexity refactor (#33), flaky `StopWatch` test fix (#31), CI docs fix |
| GitHub `main` | fleet-standard `.editorconfig`/DotSettings sync, whole-repo `jb cleanupcode` sweep, reasons on inline aislop suppressions |

Force-pushing either way would have deleted real work, so this is a merge.

## How the conflicts were resolved

All ten conflicts had the same shape: Gitea refactored code that GitHub
reformatted in place. Gitea's side won every one, because the refactor
supersedes what GitHub's version documents or restyles.

- `UsnJournalEntry` moved to an options object plus a `JournalEntryFactory`
  test helper. That removes the too-many-params problem GitHub's new
  suppression comment merely explained, so the suppression is dropped as moot.
- `BrokerFrame.cs` and `JournalBrokerScanSession.Start.cs` extracted types and
  code out of `BrokerProtocol.cs` and `JournalBrokerScanSession.cs`. Taking
  GitHub's reformatted in-place copies would have duplicated the types.
- `FreeUsnJournalInfo` / `FreeUsnJournalResult` moved to `unique_ptr` RAII, so
  the manual deletes and their `cpp-manual-delete` suppressions are both gone.

GitHub's non-conflicting hunks are preserved. That includes the redundant-cast
removal in `usn_journal.cpp`: `newCapacity` and `entryCount` are both
`uint64_t` and `sizeof` yields `size_t`, so those casts were no-ops on this
x64-only target.

The second commit extends the formatting sweep to the four files the Gitea
refactor introduced, which the original sweep predates. Running `cleanupcode`
over the merged tree touched exactly those four and nothing else, which is
also evidence that the CLT that ran it (2026.1.4) and the one that produced
the original sweep (2026.2) agree on every already-swept file.

## Verification

Verified locally on the Linux lane, the same `scripts/coverage-linux.sh` the
`linux` CI job runs: **8 native smoke tests and 281 managed tests pass, 34
skipped, 315 total**, identical to the pre-merge baseline on `main`. Native
line coverage 63.0%, managed 96.88%.

The managed percentage moves from 97.25% because the denominator grows
(1819 to 2183 coverable lines): the fleet-standard style converts
expression-bodied members to block bodies, which adds lines. No test was
removed or weakened.

**The Windows lane was not run locally and needs CI.** `MFTLibNative`'s
CMake build gates `usn/usn_journal.cpp`, `dllmain.cpp`, and
`platform_win32.cpp` behind `if(WIN32)`, so the one native file touched here
never compiled on the Linux side. The MSVC build and the admin USN tests are
exactly what the `windows` job covers.

Local `aislop ci .` improves against pristine `main` (40 warnings vs 113,
zero errors on both), though a local run does not reproduce the repo's
score-100 CI state, so treat the `quality-gate` job as the real signal.

---------

Reviewed-on: https://gitea.llamabox.sticktoitive.net/schoen/MFTLib/pulls/34

+++ MFTLibNative/mft/mft.internal.h
@@ -12,7 +12,6 @@
 #include "../internal.h"
 
 namespace mftlib::ntfs {
-
 struct DataRun {
     int64_t clusterOffset;
     uint64_t clusterCount;
@@ -40,7 +39,6 @@ bool ReadMFTRecord(HANDLE volumeHandle, const std::vector<DataRun>& mftRuns, uin
                    uint64_t recordNumber);
 #endif
 }  // namespace detail
-
 }  // namespace mftlib::ntfs
 
 // Bundles the filename-filter parameters so they travel as a single argument
@@ -72,14 +70,14 @@ struct PathLookup {
     std::atomic<uint64_t> namesDropped;
 
     bool init(uint64_t totalRecords) {
-        parents = static_cast<uint64_t*>(calloc(static_cast<size_t>(totalRecords), sizeof(uint64_t)));
-        nameLens = static_cast<uint8_t*>(calloc(static_cast<size_t>(totalRecords), sizeof(uint8_t)));
-        nameOffsets = static_cast<uint32_t*>(calloc(static_cast<size_t>(totalRecords), sizeof(uint32_t)));
+        parents = static_cast<uint64_t*>(calloc(totalRecords, sizeof(uint64_t)));
+        nameLens = static_cast<uint8_t*>(calloc(totalRecords, sizeof(uint8_t)));
+        nameOffsets = static_cast<uint32_t*>(calloc(totalRecords, sizeof(uint32_t)));
         // Each name entry can be up to 255 WCHAR units = 510 bytes; use 32 bytes avg * 2 for bytes.
         // A test hook can shrink the pool to exercise the exhaustion path.
         uint64_t capacityOverride = NamePoolCapacityOverride();
         namePoolCapacity = (capacityOverride != 0U) ? capacityOverride : totalRecords * 64;  // bytes
-        namePool = static_cast<uint8_t*>(malloc(static_cast<size_t>(namePoolCapacity)));
+        namePool = static_cast<uint8_t*>(malloc(namePoolCapacity));
         namePoolUsed = 0;
         namesDropped = 0;
         return (parents != nullptr) && (nameLens != nullptr) && (nameOffsets != nullptr) && (namePool != nullptr);
@@ -95,7 +93,7 @@ struct PathLookup {
             return;
         }
         nameOffsets[recordIndex] = static_cast<uint32_t>(offset);
-        memcpy(namePool + offset, name, static_cast<size_t>(byteCount));
+        memcpy(namePool + offset, name, byteCount);
         nameLens[recordIndex] = nameLen;
     }
 
@@ -112,7 +110,7 @@ struct PathLookup {
 struct SliceResult {
     std::vector<MftFileEntry> entries;
 
-    void init(uint64_t cap) { entries.reserve(static_cast<size_t>(cap)); }
+    void init(uint64_t cap) { entries.reserve(cap); }
 };
 
 // Read-only inputs that steer record scanning: the name filter, the optional
